requirements.txt
# Core ML and probabilistic programming
torch>=2.1.0
pyro-ppl>=1.8.6

# Data handling and preprocessing
pandas>=2.1.1
numpy>=1.26.0
scikit-learn>=1.3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Interactive development
jupyterlab>=4.0.7
tqdm>=4.66.1

# Optional: model interpretability and experiment tracking
# shap>=0.44.0
# optuna>=3.3.0
# wandb>=0.16.0
